+++ types/protocol.d.ts
@@ -9200,6 +9200,23 @@ export namespace Protocol {
             errors?: SignedExchangeError[];
         }
 
+        export type CrossOriginOpenerPolicyValue = ('SameOrigin' | 'SameOriginAllowPopups' | 'UnsafeNone' | 'SameOriginPlusCoep');
+
+        export interface CrossOriginOpenerPolicyStatus {
+            value: CrossOriginOpenerPolicyValue;
+        }
+
+        export type CrossOriginEmbedderPolicyValue = ('None' | 'RequireCorp');
+
+        export interface CrossOriginEmbedderPolicyStatus {
+            value: CrossOriginEmbedderPolicyValue;
+        }
+
+        export interface SecurityIsolationStatus {
+            coop: CrossOriginOpenerPolicyStatus;
+            coep: CrossOriginEmbedderPolicyStatus;
+        }
+
         export interface CanClearBrowserCacheResponse {
             /**
              * True if browser cache can be cleared.
@@ -9564,6 +9581,17 @@ export namespace Protocol {
              * To be sent in Sec-CH-UA-* headers and returned in navigator.userAgentData
              */
             userAgentMetadata?: Emulation.UserAgentMetadata;
+        }
+
+        export interface GetSecurityIsolationStatusRequest {
+            /**
+             * If no frameId is provided, the status of the target is provided.
+             */
+            frameId?: Page.FrameId;
+        }
+
+        export interface GetSecurityIsolationStatusResponse {
+            status: SecurityIsolationStatus;
         }
 
         /**